# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Script Master Labs \u2014 SqueezeOS, 402Proof & Ghost Layer API",
    "version": "2.0.0",
    "description": "Three-system platform: SqueezeOS (institutional AI trading intelligence), 402Proof (x402/HTTP 402 autonomous agent payment compliance firewall on XRPL), and Ghost Layer (dual-chain transparent toll gateway bridging XRPL and Base/EVM). All premium SqueezeOS endpoints follow the x402 standard \u2014 agents pay RLUSD on XRPL, no API keys required. Ghost Layer routes XRPL and Base chain USDC payments with full fee disclosure. Agent SDK: https://four02proof.onrender.com/agent/",
    "contact": {
      "name": "Script Master Labs",
      "url": "https://scriptmasterlabs.com/terminal"
    },
    "x-llm-summary": "Two-system platform: SqueezeOS delivers real-time market intelligence (squeeze scanner, options flow, IWM 0DTE, AI council verdicts). 402Proof handles x402-compliant RLUSD micropayment settlement on XRPL so autonomous agents can pay per call with no API keys. Loyalty tiers reward cumulative spend. MCP middleware available via npm @relayos/mcp-paywall."
  },
  "servers": [
    {
      "url": "https://four02proof.onrender.com",
      "description": "402Proof \u2014 Payment Settlement Server (x402 compliant, Agent Passport, compliance receipts)"
    },
    {
      "url": "https://squeezeos-terminal.vercel.app",
      "description": "SqueezeOS \u2014 Market Intelligence API (Railway, x402 gated)"
    },
    {
...
    }
  ],
  "paths": {
...
    "/api/futures/create": {
      "post": {
        "summary": "Open a Signal Futures position (free)",
        "description": "Predict the next SqueezeOS council verdict bias for a symbol and stake RLUSD. Auto-settles when verdict publishes. Zero custody — SqueezeOS tracks proof only. Winner takes 95% of pot.",
        "operationId": "futuresCreate",
        "tags": ["Futures"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creator_wallet", "symbol", "predicted_bias"],
                "properties": {
                  "creator_wallet":  {"type": "string"},
                  "symbol":          {"type": "string", "enum": ["IWM","SPY","QQQ","GME","AMC","MSTR","NVDA","TSLA","PLTR","HOOD"]},
                  "predicted_bias":  {"type": "string", "enum": ["BULLISH","BEARISH","NEUTRAL"]},
                  "stake_rlusd":     {"type": "number", "default": 0.05, "minimum": 0.01, "maximum": 50},
                  "session":         {"type": "string", "enum": ["PRE_MARKET","OPEN","MIDDAY","POWER_HOUR","CLOSE","ANY"], "default": "ANY"},
                  "ttl_hours":       {"type": "integer", "default": 8},
                  "note":            {"type": "string", "maxLength": 300}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Future created — awaiting taker"},
          "400": {"description": "Validation error"},
          "429": {"description": "Max active futures per wallet reached"}
        }
      }
    },
    "/api/futures/take/{future_id}": {
      "post": {
        "summary": "Take the opposite side of a Signal Futures position",
        "description": "Bet against the creator's predicted bias. You win if the verdict does NOT match creator's prediction.",
        "operationId": "futuresTake",
        "tags": ["Futures"],
        "parameters": [
          {"name": "future_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["taker_wallet"],
                "properties": {
                  "taker_wallet": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Position taken — stakes locked"},
          "400": {"description": "Future not OPEN or expired"},
          "404": {"description": "Future not found"}
        }
      }
    },
    "/api/futures": {
      "get": {
        "summary": "Browse open Signal Futures positions (free)",
        "operationId": "futuresList",
        "tags": ["Futures"],
        "parameters": [
          {"name": "symbol", "in": "query", "schema": {"type": "string"}},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["OPEN","ACTIVE","SETTLED","EXPIRED"]}},
          {"name": "bias",   "in": "query", "schema": {"type": "string", "enum": ["BULLISH","BEARISH","NEUTRAL"]}},
          {"name": "limit",  "in": "query", "schema": {"type": "integer", "default": 50, "maximum": 200}}
        ],
        "responses": {
          "200": {"description": "List of futures"}
        }
      }
    },
    "/api/futures/settle/{future_id}": {
      "post": {
        "summary": "Force-settle a future against latest verdict",
        "operationId": "futuresSettle",
        "tags": ["Futures"],
        "parameters": [
          {"name": "future_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Settlement proof"},
          "400": {"description": "No verdict found or future not ACTIVE"}
        }
      }
    },
    "/api/futures/leaderboard": {
      "get": {
        "summary": "Top Signal Futures predictors by wins",
        "operationId": "futuresLeaderboard",
        "tags": ["Futures"],
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20, "maximum": 100}}
        ],
        "responses": {
          "200": {"description": "Leaderboard"}
        }
      }
    },
    "/api/futures/wallet/{wallet}": {
      "get": {
        "summary": "All futures for a wallet (creator or taker)",
        "operationId": "futuresWallet",
        "tags": ["Futures"],
        "parameters": [
          {"name": "wallet", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Wallet futures history"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "ADMIN_TOKEN from Render environment. Generate with: openssl rand -hex 32. Required for /v1/admin/* endpoints."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Merchant API key obtained via POST /v1/merchant/register. Used for managing registered endpoints."
      },
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment-Token",
        "description": "JWT access token obtained via the 402Proof payment flow. Agents: hit the endpoint, handle the 402, pay RLUSD on XRPL, call /v1/verify, then retry with this header. Token TTL: 1 hour. Python SDK: pip install requests xrpl-py + use Proof402Client. NPM middleware: @relayos/mcp-paywall."
      }
    },
...
  },
...
}